+++ internal/github/client.go
@@ -3,6 +3,7 @@ package github
 import (
 	"context"
 	"fmt"
+	"net/http"
 	"os"
 	"strconv"
 
@@ -106,4 +107,24 @@ func (c *Client) GetCheck(checkName string) (*github.CheckRun, error) {
 		return nil, nil
 	}
 	return checks.CheckRuns[0], nil
+}
+
+func (c *Client) FilePresentInRepo(filepath string) (bool, error) {
+
+	_, _, resp, err := c.Repositories.GetContents(c.Ctx, owner, c.Repo, filepath, &github.RepositoryContentGetOptions{
+		Ref: c.Sha,
+	})
+
+	if err != nil {
+		if resp != nil && resp.StatusCode == http.StatusNotFound {
+			fmt.Printf("'%s' doesn't exist in `%s`\n", filepath, c.Repo)
+			return false, nil
+		} else {
+			fmt.Printf(" Error occured while checking for '%s' in `%s`\n", filepath, c.Repo)
+			return false, err
+		}
+	} else {
+		fmt.Printf("'%s' exists in `%s`\n", filepath, c.Repo)
+		return true, nil
+	}
 }