+++ CHANGELOG.md
@@ -24,3 +24,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Label overrides can be used even if repo doesn't have any config setup
 * Bumped Go to v1.20
 * Migrated image registry to ACR
+* Automatically add "E2E Test Suites" required check when ./tests/e2e/config.yaml present in repo.